Upgrade Wheels to 2.1

+++ wheels/migrator/functions.cfm
@@ -18,8 +18,8 @@ public struct function init(
 /**
  * Migrates database to a specified version. Whilst you can use this in your application, the recommended useage is via either the CLI or the provided GUI interface
  *
- * [section: Configuration]
- * [category: Database Migrations]
+ * [section: Migrator]
+ * [category: General Functions]
  *
  * @version The Database schema version to migrate to
  */
@@ -96,8 +96,8 @@ public string function migrateTo(string version="") {
 /**
  * Shortcut function to migrate to the latest version
  *
- * [section: Configuration]
- * [category: Database Migrations]
+ * [section: Migrator]
+ * [category: General Functions]
  */
 public string function migrateToLatest() {
 	local.migrations=getAvailableMigrations();
@@ -112,8 +112,8 @@ public string function migrateToLatest() {
 /**
  * Returns current database version. Whilst you can use this in your application, the recommended useage is via either the CLI or the provided GUI interface
  *
- * [section: Configuration]
- * [category: Database Migrations]
+ * [section: Migrator]
+ * [category: General Functions]
  */
 public string function getCurrentMigrationVersion() {
 	return ListLast($getVersionsPreviouslyMigrated());
@@ -122,8 +122,8 @@ public string function getCurrentMigrationVersion() {
 /**
  * Creates a migration file. Whilst you can use this in your application, the recommended useage is via either the CLI or the provided GUI interface
  *
- * [section: Configuration]
- * [category: Database Migrations]
+ * [section: Migrator]
+ * [category: General Functions]
  */
 public string function createMigration(
 	required string migrationName,
@@ -140,8 +140,8 @@ public string function createMigration(
 /**
  * Searches db/migrate folder for migrations. Whilst you can use this in your application, the recommended useage is via either the CLI or the provided GUI interface
  *
- * [section: Configuration]
- * [category: Database Migrations]
+ * [section: Migrator]
+ * [category: General Functions]
  *
  * @path Path to Migration Files: defaults to /migrator/migrations/
  */
@@ -183,13 +183,13 @@ public array function getAvailableMigrations(string path=this.paths.migrate) {
 /**
  * Reruns the specified migration version. Whilst you can use this in your application, the recommended useage is via either the CLI or the provided GUI interface
  *
- * [section: Configuration]
- * [category: Database Migrations]
+ * [section: Migrator]
+ * [category: General Functions]
  *
  * @version The Database schema version to rerun
  */
 public string function redoMigration(string version="") {
-	var currentVersion = getCurrentMigrationVersion();
+	local.currentVersion = getCurrentMigrationVersion();
 	local.appKey = $appKey();
 	if (Len(arguments.version)) {
 		currentVersion = arguments.version;
@@ -226,10 +226,11 @@ public string function redoMigration(string version="") {
  */
 private void function $setVersionAsMigrated(required string version) {
 	local.appKey = $appKey();
-	$query(
-		datasource=application[local.appKey].dataSourceName,
-		sql="INSERT INTO #application[local.appKey].migratorTableName# (version) VALUES ('#$sanitiseVersion(arguments.version)#')"
-	);
+	if(!structKeyExists(request, "$wheelsDebugSQL"))
+		$query(
+			datasource=application[local.appKey].dataSourceName,
+			sql="INSERT INTO #application[local.appKey].migratorTableName# (version) VALUES ('#$sanitiseVersion(arguments.version)#')"
+		);
 }
 
 /**
@@ -237,10 +238,11 @@ private void function $setVersionAsMigrated(required string version) {
  */
 private void function $removeVersionAsMigrated(required string version) {
 	local.appKey = $appKey();
-	$query(
-		datasource=application[local.appKey].dataSourceName,
-		sql="DELETE FROM #application[local.appKey].migratorTableName# WHERE version = '#$sanitiseVersion(arguments.version)#'"
-	);
+	if(!structKeyExists(request, "$wheelsDebugSQL"))
+		$query(
+			datasource=application[local.appKey].dataSourceName,
+			sql="DELETE FROM #application[local.appKey].migratorTableName# WHERE version = '#$sanitiseVersion(arguments.version)#'"
+		);
 }
 
 /**
@@ -287,9 +289,10 @@ private string function $copyTemplateMigrationAndRename(
 		DirectoryCreate(this.paths.migrate);
 	}
 	try {
+		local.appKey = $appKey();
 		local.templateContent = FileRead(local.templateFile);
-		if (Len(Trim(application.wheels.rootcomponentpath))) {
-			local.extendsPath = application.wheels.rootcomponentpath & ".wheels.migrator.Migration";
+		if (Len(Trim(application[local.appKey].rootcomponentpath))) {
+			local.extendsPath = application[local.appKey].rootcomponentpath & ".wheels.migrator.Migration";
 		}
 		local.templateContent = Replace(local.templateContent, "[extends]", local.extendsPath);
 		local.templateContent = Replace(local.templateContent, "[description]", Replace(arguments.migrationName, """", "&quot;", "all"));
